Update version and add Docker detection
Updated the version of Aiursoft.CSTools to 8.0.2 in the project file. Added a new function named 'IsInDocker' in EntryExtends.cs file to check if the application is running in a Docker container. This function reads the environment variables and checks for a file inherent in Docker environments to make its determination.